backend/requirements.txt
livekit-agents==1.3.1
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
livekit-plugins-turn_detector
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn
google-genai
langfuse
opentelemetry-sdk
opentelemetry-api
weaviate-client
requests>=2.31.0
asyncpg>=0.29.0
